backend track 2/requirements.txt
fastapi
uvicorn
langchain-huggingface
langchain_community
transformers==4.45.2
google-cloud-translate
google-cloud-secret-manager
sqlalchemy<2.0
psycopg2-binary
chromadb
sentence-transformers
pandas==1.5.3
torch==2.4.1
torchvision==0.19.1
tf-keras